cast

+++ Library/Sources/Stroika/Foundation/Configuration/SystemConfiguration.cpp
@@ -532,7 +532,7 @@ SystemConfiguration::Memory Configuration::GetSystemConfiguration_Memory ()
 #elif   qPlatform_POSIX
     // page size cannot change while running, but number of pages can
     // (e.g. https://pubs.vmware.com/vsphere-50/index.jsp?topic=%2Fcom.vmware.vsphere.vm_admin.doc_50%2FGUID-0B4C3128-F854-43B9-9D80-A20C0C8B0FF7.html)
-    static  const   size_t  kPageSize_   { ::sysconf (_SC_PAGESIZE) };
+    static  const   size_t  kPageSize_   { static_cast<size_t> (::sysconf (_SC_PAGESIZE)) };
     result.fPageSize = kPageSize_;
     result.fTotalPhysicalRAM = ::sysconf (_SC_PHYS_PAGES) * kPageSize_;
 #elif   qPlatform_Windows
